+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.3
+
+### Added 
+
+- Minor improvements to weblutra
+- Adding unit tests for pOTTR's stOTTR files
+
+### Removed
+
+- Revert Core Type checking instances from 0.6.2, bugs found.
+
 ## 0.6.2
 
 ### Added